+++ linstor_client/commands/rsc_dfn_cmds.py
@@ -338,6 +338,7 @@ class ResourceDefinitionCommands(Commands):
         tbl.show()
 
     def list(self, args):
+        args = self.merge_config_args('resource-definition.list', args)
         lstmsg = self._linstor.resource_dfn_list(
             query_volume_definitions=False,
             filter_by_resource_definitions=args.resource_definitions,